# Frontend/workload-management/src/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8088/api/v1",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/workload": {
      "get": {
        "tags": [
          "Workload"
        ],
        "operationId": "findAllWorkloads",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "teachingStaff.name"
            }
          },
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "asc"
            }
          },
          {
            "name": "filters",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseWorkloadResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Workload"
        ],
        "operationId": "saveWorkload",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/workload-settings": {
      "get": {
        "tags": [
          "WorkloadSettings"
        ],
        "operationId": "findAllWorkloadSettings",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkloadSettingsResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "WorkloadSettings"
        ],
        "operationId": "saveWorkloadSettings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkloadSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
...
    "/workload/{workloadId}": {
      "get": {
        "tags": [
          "Workload"
        ],
        "operationId": "findWorkloadById",
        "parameters": [
          {
            "name": "workloadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkloadResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Workload"
        ],
        "operationId": "deleteWorkloadById",
        "parameters": [
          {
            "name": "workloadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Workload"
        ],
        "operationId": "updateWorkloadById",
        "parameters": [
          {
            "name": "workloadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/workload-settings/{workloadSettingsId}": {
      "get": {
        "tags": [
          "WorkloadSettings"
        ],
        "operationId": "findWorkloadSettingsById",
        "parameters": [
          {
            "name": "workloadSettingsId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkloadSettingsResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "WorkloadSettings"
        ],
        "operationId": "deleteWorkloadSettingsById",
        "parameters": [
          {
            "name": "workloadSettingsId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "WorkloadSettings"
        ],
        "operationId": "updateWorkloadSettingsById",
        "parameters": [
          {
            "name": "workloadSettingsId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkloadSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
...
    "/workload/my-workloads": {
      "get": {
        "tags": [
          "Workload"
        ],
        "operationId": "findAllMyWorkloads",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "teachingStaff.name"
            }
          },
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "asc"
            }
          },
          {
            "name": "filters",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseObject"
                }
              }
            }
          }
        }
      }
    },
    "/workload-stats/summary": {
      "get": {
        "tags": [
          "WorkloadStats"
        ],
        "operationId": "getWorkloadSummary",
        "parameters": [
          {
            "name": "semesterId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/workload-stats/faculty-distribution": {
      "get": {
        "tags": [
          "WorkloadStats"
        ],
        "operationId": "getFacultyDistribution",
        "parameters": [
          {
            "name": "semesterId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/workload-stats/course-distribution": {
      "get": {
        "tags": [
          "WorkloadStats"
        ],
        "operationId": "getCourseDistribution",
        "parameters": [
          {
            "name": "semesterId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/workload-stats/class-distribution": {
      "get": {
        "tags": [
          "WorkloadStats"
        ],
        "operationId": "getClassDistribution",
        "parameters": [
          {
            "name": "semesterId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "WorkloadRequest": {
        "type": "object",
        "properties": {
          "teachingStaffId": {
            "type": "integer",
            "format": "int32"
          },
          "semesterId": {
            "type": "integer",
            "format": "int32"
          },
          "comments": {
            "type": "string"
          },
          "includeInBudget": {
            "type": "string"
          },
          "budgetPosition": {
            "type": "string",
            "enum": [
              "ZB",
              "PF",
              "M",
              "GRANT"
            ]
          },
          "industryCoefficient": {
            "type": "number",
            "format": "double"
          },
          "vacationMonths": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "workingMonths": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "groupAmount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "groupForSemesterId": {
            "type": "integer",
            "format": "int32"
          },
          "courseId": {
            "type": "integer",
            "format": "int32"
          },
          "academicRankId": {
            "type": "integer",
            "format": "int32"
          },
          "myClassIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "creditPointsPerGroup": {
            "type": "number",
            "format": "double",
            "minimum": 0
          }
        },
        "required": [
          "academicRankId",
          "budgetPosition",
          "comments",
          "courseId",
          "creditPointsPerGroup",
          "groupAmount",
          "groupForSemesterId",
          "includeInBudget",
          "industryCoefficient",
          "myClassIds",
          "semesterId",
          "teachingStaffId",
          "vacationMonths"
        ]
      },
      "WorkloadSettingsRequest": {
        "type": "object",
        "properties": {
          "isGlobal": {
            "type": "boolean"
          },
          "settingName": {
            "type": "string"
          },
          "visibleColumns": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isDefault": {
            "type": "boolean"
          }
        }
      },
...
      "AcademicRank": {
        "type": "object",
        "properties": {
          "academicRankId": {
            "type": "integer",
            "format": "int32"
          },
          "rankName": {
            "type": "string"
          },
          "abbreviation": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          }
        }
      },
      "AcademicRankDetails": {
        "type": "object",
        "properties": {
          "academicRankDetailsId": {
            "type": "integer",
            "format": "int32"
          },
          "cpForFullTime": {
            "type": "number",
            "format": "double"
          },
          "salary": {
            "type": "number",
            "format": "double"
          },
          "contactHoursForFullTime": {
            "type": "number",
            "format": "double"
          },
          "semester": {
            "$ref": "#/components/schemas/Semester"
          },
          "academicRank": {
            "$ref": "#/components/schemas/AcademicRank"
          },
          "deleted": {
            "type": "boolean"
          }
        }
      },
      "Course": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "integer",
            "format": "int32"
          },
          "courseCode": {
            "type": "string"
          },
          "courseName": {
            "type": "string"
          },
          "creditPoints": {
            "type": "number",
            "format": "double"
          },
          "registrationType": {
            "type": "string"
          },
          "section": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          }
        }
      },
      "Faculty": {
        "type": "object",
        "properties": {
          "facultyId": {
            "type": "integer",
            "format": "int32"
          },
          "facultyName": {
            "type": "string"
          },
          "facultyFullName": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          }
        }
      },
      "FacultyResponse": {
        "type": "object",
        "properties": {
          "facultyId": {
            "type": "integer",
            "format": "int32"
          },
          "facultyName": {
            "type": "string"
          },
          "facultyFullName": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          }
        }
      },
      "MyClassResponse": {
        "type": "object",
        "properties": {
          "classId": {
            "type": "integer",
            "format": "int32"
          },
          "classLevel": {
            "type": "integer",
            "format": "int32"
          },
          "classFaculty": {
            "$ref": "#/components/schemas/FacultyResponse"
          },
          "classProgram": {
            "type": "string"
          },
          "classLevelAndProgram": {
            "type": "string"
          },
          "degree": {
            "type": "string",
            "enum": [
              "BACHELOR",
              "MASTER",
              "DOCTORATE",
              "PROFESSIONAL_QUALIFICATION"
            ]
          },
          "deleted": {
            "type": "boolean"
          }
        }
      },
      "PageResponseWorkloadResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkloadResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "Semester": {
        "type": "object",
        "properties": {
          "semesterId": {
            "type": "integer",
            "format": "int32"
          },
          "semesterName": {
            "type": "string",
            "enum": [
              "PAVASARIS",
              "RUDENS"
            ]
          },
          "semesterYear": {
            "type": "integer",
            "format": "int32"
          },
          "deleted": {
            "type": "boolean"
          }
        }
      },
      "StatusType": {
        "type": "object",
        "properties": {
          "statusTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "statusTypeName": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          }
        }
      },
      "TeachingStaff": {
        "type": "object",
        "properties": {
          "teachingStaffId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/StatusType"
          },
          "staffFaculty": {
            "$ref": "#/components/schemas/Faculty"
          },
          "staffAcademicRank": {
            "$ref": "#/components/schemas/AcademicRank"
          },
          "positionTitle": {
            "type": "string"
          },
          "staffPhoto": {
            "type": "string"
          },
          "staffFullName": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          },
          "rankFullName": {
            "type": "string"
          }
        }
      },
      "WorkloadResponse": {
        "type": "object",
        "properties": {
          "workloadId": {
            "type": "integer",
            "format": "int32"
          },
          "teachingStaff": {
            "$ref": "#/components/schemas/TeachingStaff"
          },
          "semester": {
            "$ref": "#/components/schemas/Semester"
          },
          "comments": {
            "type": "string"
          },
          "includeInBudget": {
            "type": "string"
          },
          "budgetPosition": {
            "type": "string",
            "enum": [
              "ZB",
              "PF",
              "M",
              "GRANT"
            ]
          },
          "industryCoefficient": {
            "type": "number",
            "format": "double"
          },
          "programCoefficient": {
            "type": "number",
            "format": "double"
          },
          "vacationMonths": {
            "type": "integer",
            "format": "int32"
          },
          "expectedSalary": {
            "type": "number",
            "format": "double"
          },
          "groupAmount": {
            "type": "integer",
            "format": "int32"
          },
          "contactHours": {
            "type": "number",
            "format": "double"
          },
          "groupForSemester": {
            "$ref": "#/components/schemas/MyClassResponse"
          },
          "course": {
            "$ref": "#/components/schemas/Course"
          },
          "academicRankDetails": {
            "$ref": "#/components/schemas/AcademicRankDetails"
          },
          "myClasses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MyClassResponse"
            }
          },
          "creditPointsPerGroup": {
            "type": "number",
            "format": "double"
          },
          "totalCreditPoints": {
            "type": "number",
            "format": "double"
          },
          "cpProportionOnFullTime": {
            "type": "number",
            "format": "double"
          },
          "salaryPerMonth": {
            "type": "number",
            "format": "double"
          },
          "cpForFullTime": {
            "type": "number",
            "format": "double"
          },
          "monthSum": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "PageResponseObject": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "WorkloadSettingsResponse": {
        "type": "object",
        "properties": {
          "workloadSettingsId": {
            "type": "integer",
            "format": "int32"
          },
          "settingName": {
            "type": "string"
          },
          "visibleColumns": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "default": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}